requirements.txt
streamlit==1.32.2
streamlit-chat==0.1.1
google-generativeai==0.3.2
toml==0.10.2 